+++ tests/testthat/test-clustree-overlay.R
@@ -7,6 +7,13 @@ iris_clusts4 <- iris_clusts
 iris_clusts4$L1 <- iris_clusts4$K1
 iris_clusts4$L2 <- iris_clusts4$K2
 
+iris_clusts5 <- iris_clusts
+iris_clusts5$L0.2 <- iris_clusts$K1
+iris_clusts5$L0.4 <- iris_clusts$K2
+iris_clusts5$L0.6 <- iris_clusts$K3
+iris_clusts5$L0.8 <- iris_clusts$K4
+iris_clusts5$L1.0 <- iris_clusts$K5
+
 seurat_clusters2 <- sc_example$seurat_clusters
 seurat_clusters2$resX <- "X"
 seurat_clusters2$TSNE1 <- sc_example$tsne[, 1]
@@ -149,4 +156,14 @@ test_that("prefix selection doesn't match wildcards", {
     expect_is(clustree_overlay(seurat_clusters2, prefix = "res.",
                                x_value = "TSNE1", y_value = "TSNE2"),
               c("gg", "ggplot"))
+})
+
+test_that("point colour works with rounded resolutions", {
+    overlay_list <- clustree_overlay(iris_clusts5, prefix = "L",
+                                     x_value = "PC1", y_value = "PC2",
+                                     plot_sides = TRUE, use_colour = "points")
+
+    expect_is(overlay_list$overlay, c("gg", "ggplot"))
+    expect_is(overlay_list$x_side, c("gg", "ggplot"))
+    expect_is(overlay_list$y_side, c("gg", "ggplot"))
 })